net:qos: police action offloading parameter 'burst' change to the original value
Since 'tcfp_burst' with TICK factor, driver side always need to recover
it to the original value, this patch moves the generic calculation and
recover to the 'burst' original value before offloading to device driver.

Backported upstream commit
5f035af76e51 ("net:qos: police action offloading parameter 'burst' change
to the original value")

Acked-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/48867
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>

Orabug: 33889959

(cherry picked from commit 95989cc7748958b5ee97e4985e86e5838bdd862d)

+++ include/net/flow_offload.h
@@ -190,7 +190,7 @@ struct flow_action_entry {
 			bool			truncate;
 		} sample;
 		struct {				/* FLOW_ACTION_POLICE */
-			s64			burst;
+			u32			burst;
 			u64			rate_bytes_ps;
 		} police;
 		struct {				/* FLOW_ACTION_CT */
